+++ dc-model-jackson/src/main/java/de/digitalcollections/model/jackson/DigitalCollectionsModelModule.java
@@ -42,7 +42,10 @@ import de.digitalcollections.model.identifiable.entity.manifestation.ProductionI
 import de.digitalcollections.model.identifiable.entity.manifestation.PublicationInfo;
 import de.digitalcollections.model.identifiable.entity.manifestation.Publisher;
 import de.digitalcollections.model.identifiable.entity.relation.EntityRelation;
+import de.digitalcollections.model.identifiable.entity.relation.EntityToFileResourceRelation;
 import de.digitalcollections.model.identifiable.entity.work.Work;
+import de.digitalcollections.model.identifiable.relation.IdentifiableToEntityRelation;
+import de.digitalcollections.model.identifiable.relation.IdentifiableToFileResourceRelation;
 import de.digitalcollections.model.identifiable.resource.ApplicationFileResource;
 import de.digitalcollections.model.identifiable.resource.AudioFileResource;
 import de.digitalcollections.model.identifiable.resource.FileResource;
@@ -85,7 +88,10 @@ import de.digitalcollections.model.jackson.mixin.identifiable.entity.manifestati
 import de.digitalcollections.model.jackson.mixin.identifiable.entity.manifestation.PublicationInfoMixIn;
 import de.digitalcollections.model.jackson.mixin.identifiable.entity.manifestation.PublisherMixIn;
 import de.digitalcollections.model.jackson.mixin.identifiable.entity.relation.EntityRelationMixIn;
+import de.digitalcollections.model.jackson.mixin.identifiable.entity.relation.EntityToFileResourceRelationMixIn;
 import de.digitalcollections.model.jackson.mixin.identifiable.entity.work.WorkMixIn;
+import de.digitalcollections.model.jackson.mixin.identifiable.relation.IdentifiableToEntityRelationMixIn;
+import de.digitalcollections.model.jackson.mixin.identifiable.relation.IdentifiableToFileResourceRelationMixIn;
 import de.digitalcollections.model.jackson.mixin.identifiable.resource.ApplicationFileResourceMixIn;
 import de.digitalcollections.model.jackson.mixin.identifiable.resource.AudioFileResourceMixIn;
 import de.digitalcollections.model.jackson.mixin.identifiable.resource.FileResourceMixIn;
@@ -284,12 +290,18 @@ public class DigitalCollectionsModelModule extends SimpleModule {
     context.setMixInAnnotations(DigitalObject.class, DigitalObjectMixIn.class);
     context.setMixInAnnotations(Entity.class, EntityMixIn.class);
     context.setMixInAnnotations(EntityRelation.class, EntityRelationMixIn.class);
+    context.setMixInAnnotations(
+        EntityToFileResourceRelation.class, EntityToFileResourceRelationMixIn.class);
     context.setMixInAnnotations(Event.class, EventMixIn.class);
     context.setMixInAnnotations(FileResource.class, FileResourceMixIn.class);
     context.setMixInAnnotations(FilterCriterion.class, FilterCriterionMixIn.class);
     context.setMixInAnnotations(Filtering.class, FilteringMixIn.class);
     context.setMixInAnnotations(Heading.class, HeadingMixIn.class);
     context.setMixInAnnotations(Identifiable.class, IdentifiableMixIn.class);
+    context.setMixInAnnotations(
+        IdentifiableToEntityRelation.class, IdentifiableToEntityRelationMixIn.class);
+    context.setMixInAnnotations(
+        IdentifiableToFileResourceRelation.class, IdentifiableToFileResourceRelationMixIn.class);
     context.setMixInAnnotations(Identifier.class, IdentifierMixIn.class);
     context.setMixInAnnotations(IdentifierType.class, IdentifierTypeMixIn.class);
     context.setMixInAnnotations(IFrame.class, IFrameMixIn.class);